requirements.txt
requests==2.28.1
python-twitter==3.5
unicodecsv==0.14.1
pyinstaller==5.6.2